fix: establish consistency
We have durable objects, services (classes) and utils (functions/helpers). Naming is much clearer too now.
