+++ src/durable-objects/bns-do.ts
@@ -2,9 +2,9 @@ import { DurableObject } from 'cloudflare:workers';
 import { Env } from '../../worker-configuration';
 import { AppConfig } from '../config';
 import { validateStacksAddress } from '@stacks/transactions';
-import { createJsonResponse } from '../utils/requests-responses';
-import { getKnownAddresses } from '../utils/address-store';
-import { getNameFromAddress, initStacksFetcher } from '../utils/bns-v2';
+import { createJsonResponse } from '../utils/requests-responses-util';
+import { getKnownAddresses } from '../utils/address-store-util';
+import { getNameFromAddress, initStacksFetcher } from '../utils/bns-v2-util';
 
 /**
  * Durable Object class for the BNS API
