+++ src/durable-objects/stx-city-do.ts
@@ -1,8 +1,8 @@
 import { DurableObject } from 'cloudflare:workers';
 import { Env } from '../../worker-configuration';
 import { AppConfig } from '../config';
-import { ApiRateLimiterService } from '../rate-limiter';
-import { createJsonResponse } from '../utils/requests-responses';
+import { ApiRateLimiterService } from '../services/api-rate-limiter-service';
+import { createJsonResponse } from '../utils/requests-responses-util';
 
 type Metrics = {
 	price_usd: number;
