+++ src/index.ts
@@ -5,7 +5,7 @@ import { HiroApiDO } from './durable-objects/hiro-api-do';
 import { StxCityDO } from './durable-objects/stx-city-do';
 import { SupabaseDO } from './durable-objects/supabase-do';
 import { ContractCallsDO } from './durable-objects/contract-calls-do';
-import { corsHeaders, createJsonResponse } from './utils/requests-responses';
+import { corsHeaders, createJsonResponse } from './utils/requests-responses-util';
 
 // export the Durable Object classes we're using
 export { BnsApiDO, HiroApiDO, StxCityDO, SupabaseDO, ContractCallsDO };
