+++ src/utils/bns-v2-util.ts
@@ -2,7 +2,7 @@ import { Env } from '../../worker-configuration';
 import { AppConfig } from '../config';
 import { getFetchOptions, setFetchOptions } from '@stacks/common';
 import { BufferCV, ClarityType, principalCV, TupleCV } from '@stacks/transactions';
-import { StacksContractFetcher } from '../services/stacks-contract-fetcher';
+import { StacksContractFetcher } from '../services/stacks-contract-data-service';
 
 const BNS_CONTRACT_ADDRESS = 'SP2QEZ06AGJ3RKJPBV14SY1V5BBFNAW33D96YPGZF';
 const BNS_CONTRACT_NAME = 'BNS-V2';
